refactor(breach-history): extract BreachTable + BreachRow + filter helpers
Move BreachRow (row math + JSX), BreachTable (thead/tbody scaffold), and
filter helpers (whereForBucket, composeWhere, label maps) into
breach-history/ sub-directory. breach-history-panel.tsx shrinks from 672
to ~390 lines. All 51 characterization tests + 1691 suite tests pass.
